refactor: Common Enum Package #6
