+++ src/main/java/sopt/comfit/user/domain/User.java
@@ -6,6 +6,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import sopt.comfit.global.base.BaseTimeEntity;
+import sopt.comfit.global.enums.EIndustry;
 import sopt.comfit.university.domain.University;
 
 @Entity
